refactor(cli): import kernel through barrel
