+++ packages/cli/src/cli/parse-args.ts
@@ -3,7 +3,7 @@ import { cliError, CliErrorCode } from "./error-codes.ts";
 import { applyJsonInputLayer } from "./json-input.ts";
 import { parseRegisteredCommand } from "./parser-registry.ts";
 import { stripGlobalOptions } from "./parse-options.ts";
-import type { HarnessLayoutOverrides } from "../../../kernel/src/layout/index.ts";
+import type { HarnessLayoutOverrides } from "../../../kernel/src/index.ts";
 import type { CliResult, ParsedCommand } from "./types.ts";
 
 export function parseArgs(argv: ReadonlyArray<string>): { readonly ok: true; readonly value: ParsedCommand } | { readonly ok: false; readonly error: CliResult["error"] } {
