+++ packages/cli/src/cli/parsers/core-task.ts
@@ -1,5 +1,5 @@
-import { domainStatuses, isDomainStatus } from "../../../../kernel/src/domain/index.ts";
-import { slugifyTaskTitle } from "../../../../kernel/src/layout/index.ts";
+import { domainStatuses, isDomainStatus } from "../../../../kernel/src/index.ts";
+import { slugifyTaskTitle } from "../../../../kernel/src/index.ts";
 import { cliError, CliErrorCode } from "../error-codes.ts";
 import { readOption, readRequiredValueOption } from "../parse-options.ts";
 import type { CliResult, ParsedCommand } from "../types.ts";
