+++ packages/cli/src/commands/adopt.ts
@@ -2,9 +2,9 @@ import { Effect } from "effect";
 import path from "node:path";
 import { makeLocalWriteCoordinator } from "../../../adapters/local/src/index.ts";
 import { makeMulticaAdoptionService, makeMulticaLifecycleEngine, type MulticaClient, type MulticaRawIssue } from "../../../adapters/multica/src/index.ts";
-import type { ArtifactStoreError, EngineError, ExternalRef, TaskId, WriteError } from "../../../kernel/src/domain/index.ts";
-import type { HarnessLayoutInput } from "../../../kernel/src/layout/index.ts";
-import { resolveHarnessLayout, taskPackagePath } from "../../../kernel/src/layout/index.ts";
+import type { ArtifactStoreError, EngineError, ExternalRef, TaskId, WriteError } from "../../../kernel/src/index.ts";
+import type { HarnessLayoutInput } from "../../../kernel/src/index.ts";
+import { resolveHarnessLayout, taskPackagePath } from "../../../kernel/src/index.ts";
 import type { CliResult } from "../cli/types.ts";
 
 export interface AdoptMulticaAction {
