+++ packages/cli/src/commands/core/decision-relate.ts
@@ -3,8 +3,8 @@ import path from "node:path";
 import { Effect } from "effect";
 import { readDecisionDocument, type DecisionWriteService, type DecisionWriteRejected } from "../../../../application/src/index.ts";
 import { deriveRelationId, type DecisionPackage, type EntityRelationRecord, type WriteError } from "../../../../kernel/src/index.ts";
-import { resolveHarnessLayout, taskDocumentPath, type HarnessLayoutInput } from "../../../../kernel/src/layout/index.ts";
-import { readFrontmatter, readScalar } from "../../../../kernel/src/markdown/frontmatter.ts";
+import { resolveHarnessLayout, taskDocumentPath, type HarnessLayoutInput } from "../../../../kernel/src/index.ts";
+import { readFrontmatter, readScalar } from "../../../../kernel/src/index.ts";
 import { cliError, CliErrorCode } from "../../cli/error-codes.ts";
 import type { CommandRunnerContext } from "../../cli/runner-registry.ts";
 import type { CliResult, ParsedCommand } from "../../cli/types.ts";
