+++ packages/cli/src/commands/core/fact.ts
@@ -3,7 +3,7 @@ import path from "node:path";
 import { Effect } from "effect";
 import type { FactWriteRejected } from "../../../../application/src/index.ts";
 import { parseFactFlowRecords, type FactRecord, type WriteError } from "../../../../kernel/src/index.ts";
-import { resolveHarnessLayout, type HarnessLayoutInput } from "../../../../kernel/src/layout/index.ts";
+import { resolveHarnessLayout, type HarnessLayoutInput } from "../../../../kernel/src/index.ts";
 import { cliError, CliErrorCode } from "../../cli/error-codes.ts";
 import type { CommandRunner } from "../../cli/runner-registry.ts";
 import type { CliResult, ParsedCommand } from "../../cli/types.ts";
