+++ packages/cli/src/commands/core/provenance-backfill.ts
@@ -2,11 +2,11 @@ import { readFileSync } from "node:fs";
 import path from "node:path";
 import { Effect, Schema } from "effect";
 import type { CurrentSessionRef } from "../../../../kernel/src/index.ts";
-import type { TaskId, WriteError } from "../../../../kernel/src/domain/index.ts";
-import { stablePayloadHash } from "../../../../kernel/src/integrity/stable-hash.ts";
-import type { HarnessLayoutInput } from "../../../../kernel/src/layout/index.ts";
-import { listTaskIndexPaths, readFrontmatter, readScalar, resolveHarnessLayout } from "../../../../kernel/src/layout/index.ts";
-import { ProvenanceEntrySchema } from "../../../../kernel/src/schemas/common.ts";
+import type { TaskId, WriteError } from "../../../../kernel/src/index.ts";
+import { stablePayloadHash } from "../../../../kernel/src/index.ts";
+import type { HarnessLayoutInput } from "../../../../kernel/src/index.ts";
+import { listTaskIndexPaths, readFrontmatter, readScalar, resolveHarnessLayout } from "../../../../kernel/src/index.ts";
+import { ProvenanceEntrySchema } from "../../../../kernel/src/index.ts";
 import { writeCoordinatedTaskDocuments } from "../../../../kernel/src/write-coordination/write-helpers.ts";
 import type { CommandRunnerContext } from "../../cli/runner-registry.ts";
 import type { CliResult, ParsedCommand } from "../../cli/types.ts";
